LPS-97283 - extract MockLiferayPortletRequest
