+++ modules/apps/blogs/blogs-web-test/src/testIntegration/java/com/liferay/blogs/web/test/MockLiferayPortletRequest.java
@@ -1,0 +1,136 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.blogs.web.test;
+
+import com.liferay.portal.kernel.model.Portlet;
+import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+
+import java.util.Collection;
+import java.util.Map;
+
+import javax.portlet.ActionParameters;
+import javax.portlet.PortletConfig;
+import javax.portlet.PortletContext;
+import javax.portlet.PortletResponse;
+import javax.portlet.RenderParameters;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.Part;
+
+import org.springframework.mock.web.portlet.MockActionRequest;
+
+/**
+ * @author Alicia García
+ */
+public class MockLiferayPortletRequest
+	extends MockActionRequest implements LiferayPortletRequest {
+
+	@Override
+	public void cleanUp() {
+	}
+
+	@Override
+	public Map<String, String[]> clearRenderParameters() {
+		return null;
+	}
+
+	@Override
+	public void defineObjects(
+		PortletConfig portletConfig, PortletResponse portletResponse) {
+	}
+
+	@Override
+	public ActionParameters getActionParameters() {
+		return null;
+	}
+
+	@Override
+	public long getContentLengthLong() {
+		return 0;
+	}
+
+	@Override
+	public HttpServletRequest getHttpServletRequest() {
+		return null;
+	}
+
+	@Override
+	public String getLifecycle() {
+		return null;
+	}
+
+	@Override
+	public HttpServletRequest getOriginalHttpServletRequest() {
+		return null;
+	}
+
+	@Override
+	public Part getPart(String name) {
+		return null;
+	}
+
+	@Override
+	public Collection<Part> getParts() {
+		return null;
+	}
+
+	@Override
+	public long getPlid() {
+		return 0;
+	}
+
+	@Override
+	public Portlet getPortlet() {
+		return null;
+	}
+
+	@Override
+	public PortletContext getPortletContext() {
+		return null;
+	}
+
+	@Override
+	public String getPortletName() {
+		return null;
+	}
+
+	@Override
+	public HttpServletRequest getPortletRequestDispatcherRequest() {
+		return null;
+	}
+
+	@Override
+	public RenderParameters getRenderParameters() {
+		return null;
+	}
+
+	@Override
+	public String getUserAgent() {
+		return null;
+	}
+
+	@Override
+	public void invalidateSession() {
+	}
+
+	@Override
+	public void setPortletRequestDispatcherRequest(
+		HttpServletRequest httpServletRequest) {
+	}
+
+	protected long entryId;
+	protected HttpServletRequest httpServletRequest;
+
+}
